Overhaul
Dropped the procedural design for an OOP approach, data strucure is now centered around line buffers.
TAB & other special character handling is closer to working 100% of the time due to the use of a world offset variable that calculates position from the left of the screen to current/next position on the line.
Word wrapping is implemented with virtual line objects that mimic the way the console naturally draws lines.